+++ DiabloUI/src/diabedit.cpp
@@ -248,7 +248,7 @@ void __fastcall DiabEdit_SetRestrictTimer(HWND hWnd)
 {
     unsigned char* v2; // eax
 
-    SDlgSetTimer((int)hWnd, 1, 500, 0);
+    SDlgSetTimer(hWnd, 1, 500, 0);
     SetPropA(hWnd, "CURSOR", 0);
     v2 = (unsigned char*)SMemAlloc(0x100u, "C:\\Src\\Diablo\\DiabloUI\\DiabEdit.cpp", 185, 0);
     *v2 = 0;
@@ -260,7 +260,7 @@ void __fastcall DiabEdit_RemoveAllProps(HWND hWnd)
 {
     HANDLE v2; // eax
 
-    SDlgKillTimer((int)hWnd, 1);
+    SDlgKillTimer(hWnd, 1);
     RemovePropA(hWnd, "LIMIT");
     RemovePropA(hWnd, "CURSOR");
     v2 = RemovePropA(hWnd, "RESTRICTED");
